+++ code/src/cljs/sixsq/nuvla/ui/edge/spec.cljs
@@ -9,6 +9,8 @@
 
 (s/def ::nuvlaboxes any?)
 
+(s/def ::nuvlabox-locations any?)
+
 (s/def ::nuvlabox-cluster-summary any?)
 
 (s/def ::nuvlaboxes-summary any?)