+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
   - Support for OpenStack infrastructures and credentials
 
+### Changed
+
+  - NuvlaBox map view now show all NuvlaBoxes, including NBs with inferred location
+
 ## [2.22.0] - 2021-09-14
 
 ### Added